Review/parkerklein/youngho

+++ index.js
@@ -119,20 +119,30 @@ function add(fileName) {
     }
     var path1 = argv.directory1 + '/' + fileName;
     var path2 = argv.directory2 + '/' + fileName;
-    var handler1 = sync.getHandler(path1);
-    var handler2 = sync.getHandler(path2);
-    try {
-        // adds file to both directories
-        handler1.writeFile(path1, 'new File', function(){});
-        handler2.writeFile(path1, 'new File', function(){});
-        lastUpdate = new Date();
-        console.log('Files added on ' + formatTime(lastUpdate));
-    } catch (err) {
-        console.log('Failed to add new file ' + fileName);
-        console.log(err.message);
-        return;
-    }
-    console.log('Adding ' + fileName);
+
+	var handler1 = sync.getHandler(path1);
+	var handler2 = sync.getHandler(path2);
+
+	//fixed here (youngho)
+	var checkPath1 = './' + path1.split('//')[1];
+    var checkPath2 = './' + path2.split('//')[1];
+
+	try {
+		if(fs.existsSync(checkPath1) || fs.existsSync(checkPath2)){
+			console.log("file already exists so will not create new file.");
+		}else{
+       		// adds file to both directories
+       		handler1.writeFile(path1, 'new File', function(){});
+			handler2.writeFile(path2, 'new File', function(){});
+       		lastUpdate = new Date();
+       		console.log('Files added on ' + formatTime(lastUpdate));
+			console.log('Adding ' + fileName);
+		}
+   	} catch (err) {
+       	console.log('Failed to add new file ' + fileName);
+		console.log(err.message);
+       	return;
+   	}
 }
 
 // returns the last time dropbox was updated
@@ -159,7 +169,7 @@ function formatTime(time) {
     var minute = time.getMinutes();
     var second = time.getSeconds();
 
-    var update = dayName + ' ' + monthName + ' ' + day + ', ' + year + ' at ' + hour + ':' + minute + '.' + second;
+    var update = dayName + ' ' + monthName + ' ' + day + ', ' + year + ' at ' + hour + ':' + minute + ':' + second;
 
     return update;
 }
